Refactor NetworkInterfacesPrinter: Move its func getifaddrs_checked to Interfaces::update()
